+++ apps/docs/src/content/03-components/content/icon/examples/decorative.tsx
@@ -55,8 +55,11 @@ import {
   IconUser,
   IconVhost,
   IconVolume,
+  IconPath,
   Label,
   LabeledValue,
+  IconDnsZone,
+  IconEmailArchive,
 } from "@mittwald/flow-react-components";
 
 <ColumnLayout l={[1, 1, 1, 1, 1]} m={[1, 1, 1]} s={[1, 1]}>
@@ -101,6 +104,10 @@ import {
     <IconDeliveryBox />
   </LabeledValue>
   <LabeledValue>
+    <Label>EmailArchive</Label>
+    <IconEmailArchive />
+  </LabeledValue>
+  <LabeledValue>
     <Label>App</Label>
     <IconApp />
   </LabeledValue>
@@ -117,8 +124,16 @@ import {
     <IconSubdomain />
   </LabeledValue>
   <LabeledValue>
+    <Label>Domainpfad</Label>
+    <IconPath />
+  </LabeledValue>
+  <LabeledValue>
     <Label>Certificate</Label>
     <IconCertificate />
+  </LabeledValue>
+  <LabeledValue>
+    <Label>DnsZone</Label>
+    <IconDnsZone />
   </LabeledValue>
   <LabeledValue>
     <Label>Move</Label>