+++ packages/components/src/components/ComboBox/stories/Default.stories.tsx
@@ -40,6 +40,8 @@ export const Default: Story = {};
 
 export const Disabled: Story = { args: { isDisabled: true } };
 
+export const ReadOnly: Story = { args: { isReadOnly: true } };
+
 export const Required: Story = {
   args: { isRequired: true },
 };