+++ packages/components/src/components/PasswordCreationField/index.ts
@@ -1,4 +1,3 @@
-export * from "@/components/PasswordCreationField/lib/generatePasswordCreationFieldValidation";
 export * from "./view";
 
 export {